mv selector and matcher to query; separate parsing
