+++ src/query/matcher_try_from.rs
@@ -1,6 +1,6 @@
-use crate::query::matcher::{AnyVariant, Matcher};
 use crate::query::strings::{ParsedString, ParsedStringMode};
 use crate::query::{DetachedSpan, Pair, ParseError};
+use crate::select::matcher::{AnyVariant, Matcher};
 use regex::Regex;
 
 impl TryFrom<Option<Pair<'_>>> for Matcher {
