+++ src/query/mod.rs
@@ -1,18 +1,16 @@
 mod query;
-mod selectors;
 
 mod error;
-mod matcher;
 mod matcher_try_from;
 mod selector_try_from;
 mod strings;
 mod traversal;
 mod traversal_composites;
 
+pub use crate::select::matcher::*;
+pub use crate::select::selector::*;
 pub use error::*;
-pub use matcher::*;
 #[cfg(test)]
 pub use query::StringVariant;
 #[allow(unused_imports)]
 pub use query::{Error, Pair, Pairs, Query};
-pub use selectors::*;
