+++ src/select/mod.rs
@@ -1,10 +1,12 @@
 mod api;
 mod match_selector;
+pub mod matcher;
 mod sel_code_block;
 mod sel_link_like;
 mod sel_list_item;
 mod sel_section;
 mod sel_single_matcher;
 mod sel_table;
+pub mod selector;
 
 pub use api::*;
